Introduced `FastBlockLocation` for reuse in future code
